Type area/stage/map IDs and drive CityMapID from the queued stage.
Add AreaIndex; use StageIndex/MapIndex on catalog fields and XML
tables. Resolve map via STAGELIST onto HQ_PartyEnqueue → MQ_CreateGame.
CreateDevGame still hardcodes deathmatch. Align game-outbound names
(nResult, userId, CityMapID). Rename SN_UpdateEntrySystem nested row
to StageInfo (PST_STAGE_INFO).

+++ src/servers/hub/channel.cpp
@@ -520,7 +520,7 @@ void HubPacketHandler::HandlePacket_CQ_RequestAreaPopularity(ClientHandle client
 		return;
 	}
 	const Cl::CQ_RequestAreaPopularity& req = SafeCast<Cl::CQ_RequestAreaPopularity>(packetData, packetSize);
-	NT_LOG("[client%x] Client :: CQ_RequestAreaPopularity :: { area=%u }", clientHd, req.areaID);
+	NT_LOG("[client%x] Client :: CQ_RequestAreaPopularity :: { area=%d }", clientHd, (i32)req.areaID);
 
 	replication->SendAreaPopularity(clientHd, req.areaID);
 }
